amend: get function
