+++ test/index.spec.js
@@ -1,4 +1,5 @@
 const expect = require('chai').expect;
+const Joi = require('joi-strict');
 const { describe } = require('node-tdd');
 const sfs = require('smart-fs');
 const chai = require('chai');
@@ -67,5 +68,32 @@ describe('Testing index', {
       ]
     );
     expect(index.index.versions.persist(dir)).to.equal(false);
+  });
+
+  it('Testing load', async ({ dir }) => {
+    expect(index.index.versions.persist(dir)).to.equal(true);
+    expect(index.index.versions.load(dir)).to.equal(undefined);
+  });
+
+  it('Testing get', ({ dir }) => {
+    expect(index.index.versions.persist(dir)).to.equal(true);
+    index.index.versions.load(dir);
+    const result = index.index.versions.get(dir);
+    const schema = Joi.object().pattern(
+      Joi.string().valid(
+        'a2066a68e07cc088f3fb8921ba0fa4f3541b569a',
+        '127f07825e9279eb9f3bf334e5dd575916f09128',
+        '6a1b8f491e156e356ab57e8df046b9f449acb440'
+      ),
+      Joi.object().keys({
+        timestamp: Joi.number().integer(),
+        specs: Joi.object(),
+        mapping: Joi.object(),
+        fields: Joi.array().items(Joi.string()),
+        rels: Joi.object()
+      })
+    );
+    expect(Object.keys(result).sort()).to.deep.equal(['address', 'location', 'offer']);
+    expect(Object.values(result).every((e) => Joi.test(e, schema))).to.equal(true);
   });
 });
